+++ client/src/App.jsx
@@ -1,76 +1,32 @@
-import React, { useEffect, useLayoutEffect } from "react";
-import { useNavigate, useParams } from "react-router-dom";
-import cookies from 'js-cookie';
-import { useTranslation } from 'react-i18next';
+import React, { Suspense } from "react";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import './components/Languages/i18nextConf.js'
+import GoogleAnalyticsInitialize from './util/GoogleAnalytics/GoogleAnalyticsInitialize'
 
-import { pageList } from "./resource/pages"
-import getLink, { getType } from "./resource/link";
-import { languages } from "./components/Languages/Language";
+import { RedirectApp, RedirectToLink } from "./RedirectApp";
+import Vcard from "./components/vcard/Vcard";
+import Loading from "./components/Loading/Loading";
 
-
-const pages = pageList.reduce((urls, page) => {
-  urls.push(page.url);
-  return urls;
-}, [])
-
-export function RedirectApp() {
-  const { t } = useTranslation();
-  const navigate = useNavigate();
-  const redirect = useParams();
-  const currentLanguageCode = cookies.get('i18next') || 'en';
-
-  useEffect(() => {
-    if (
-      typeof redirect === "undefined" || typeof redirect.lang === "undefined" || (
-        !pages.includes(redirect.lang) &&
-        !languages.includes(redirect.lang)
-      )
-    ) {
-      navigate(`/${currentLanguageCode}`, { replace: true });
-    }
-    if (pages.includes(redirect.lang)) {
-      navigate(`/${currentLanguageCode}/${redirect.lang}`, { replace: true });
-    }
-    return () => {
-
-    }
-  }, []);
-
-  useLayoutEffect(() => {
-    if (currentLanguageCode !== redirect.lang) {
-      if (redirect.page) {
-        navigate(`/${currentLanguageCode}/${redirect.page}`, { replace: true });
-      } else {
-        navigate(`/${currentLanguageCode}`, { replace: true });
-      }
-    }
-    return () => {
-
-    }
-  }, [t])
-
-  if (redirect.page === undefined) {
-    redirect.page = "";
-  }
-  if (pages.includes(redirect.page)) {
-    const PageComponent = pageList.find((page) => page.url === redirect.page).Component;
-    return <PageComponent redirect={redirect} />
-  }
-  return "404";
-}
-
-export function RedirectToLink() {
-  const redirect = useParams();
-  const from = getType(redirect.from);
-  const to = getLink(redirect.to);
-
-
-  window.location.replace(to);
+export function App() {
 
 
   return (
-    <></>
+    <Suspense fallback={<></>}>
+      <GoogleAnalyticsInitialize />
+      <Loading />
+      <React.StrictMode>
+        <BrowserRouter>
+          <Routes>
+            <Route path="/" element={<RedirectApp />} />
+            <Route path="/:lang" element={<RedirectApp />} />
+            <Route path="/:lang/:page" element={<RedirectApp />} />
+            <Route exact path="/vcard/:type" element={<Vcard />} />
+            <Route exact path="/redirect/:from/:to" element={<RedirectToLink />} />
+          </Routes>
+        </BrowserRouter>
+      </React.StrictMode>
+    </Suspense>
   )
 }
 
-export default RedirectApp;
+export default App;
